Rename subroutine back to CHOX (from CHOX_2_8).
Note that, for some reason, this file is around half the size if the file in the last commit.
No idea why this i s. This macro version has been tested and verified by BR and so should be considered
to be the latest version. However, if any problems occur then please consider reverting to the previous version.

	modified:   Configuration/macro/iDAS_CHOX_ERAC_v2.8.xla

Change-Id: I0d8169834f607d30905d8d87390a2c738b15469d